requirements.txt
fastapi
uvicorn
pdf2image
Pillow
PyMuPDF
